Fix test fixture
